added all the forms, urls and views; almost done with the templates.
Added all the forms and views and started working on the templates, only the ones regarding the profile remain empty. Configured the urls and removed the signal for slug creation for the "lecture" model, since it's better to just overwrite the "save" method, as it is done in the "question" model.