added all templates with functioning views and forms, is_answered on question, mailjet_api integrated, requirements.txt
Added all the needed templates for the project, together with the respective views and forms. Integrated the mailjet_rest api to send welcome emails upon registration (accounts/utils and signals). Added a requirements.txt file. Added a "is_answered" boolean field to the "Question" model. Added a default profile picture and a background one, together with styling the templates.